requirements.txt
setuptools~=65.5.1
mysql-connector-python~=8.1.0
flask~=2.3.3